That kernel that I erased. I am going to have bring it back after all. The thing I wrote about yesterday would not work because it would not be an associative operation. Fuck.
Ah, this is going to take a while it seems. For real. I'll make a CPS'd version of it.

This is so annoying. There are just so many design choices here. I cannot generalize it properly.

Well, let me take a break for a bit since I just woke up and then I will start this thing.