``` let kernel10 =     "kernel10",[cuda_modules],"Does the init_redo_redo kernel work?",     """ inb s = CudaModules (1024*1024)
inl c,b,a = 6,12,256
inl a1 = s.CudaRandom.create {dst=.Normal; stddev=1f32; mean=0f32} {elem_type=float32; dim=c,b,a}
s.CudaTensor.print a1
inl o1 = s.CudaTensor.create {elem_type=float32,int64; dim=c}
inl _ =
    inl a1,o1 = CudaAux.to_dev_tensor (a1,o1)
    s.CudaKernel.init_redo_redo (inl c' k ->
        inl a1 = a1 c'
        k {
            neutral_elem=-infinityf32,-1
            redo=inl a b -> if fst a > fst b then a else b
            init=inl b k ->
                inl a1 = a1 b
                k {
                    neutral_elem=0f32
                    redo=(+)
                    init=inl a -> a1 a .get
                    }
            }
        |> inl a,b -> a / c', b
        |> o1 c' .set
        )
s.CudaTensor.print o1
    """
```

Here is the test. I think that this is actually a pretty good way of doing it. I get most of the advantages of lexical scope here and the thing does act like a mini-DSL.

1:15pm.

```
/// Maps the input and then reduces twice, first along the inner dimension and then along the middle.
met init_redo_redo w {dim=c,b,a; init} =
    assert (dim_in_a = out_a) "Input's outermost and output's dimension must be equal."
    inl x = lit_min 1024 (length a)
    inl y = lit_min (1024 / x) (length b)
    w.run {
        blockDim={y x}
        gridDim={z=min 64 (length c)}
        kernel = cuda
            inl grid_for = grid_for {blockDim gridDim}
            grid_for .z c {body=inl {i} ->
                init i <| inl {neutral_elem redo init outit} ->
                    inl x =
                        grid_for .y b {state=dyn neutral_elem; body=inl {state i} ->
                            init i <| inl {neutral_elem redo init} ->
                                grid_for .x a {state=dyn neutral_elem; body=inl {state i} -> redo state (init i)}
                                |> block_reduce_2d (blockDim.y, threadIdx.y) (blockDim.x, threadIdx.x) redo
                                |> redo state
                            }
                    if threadIdx.x = 0 then
                        inl x = block_reduce_1d (blockDim.y, threadIdx.y) redo x
                        if threadIdx.y = 0 then outit x
                    }
        }
```

I think this should work fine. Adapting these kernels for multiple functions would be difficult. No way am I going to be able to generalize it much more than this. Even this way of doing it has a disadvantage of making inference a lot more diffiuclt.

But if I am ever using this kernel it will be in some specialized function so I guess I do not have to worry about making allocators and hence inference. This is quite good as it is.

Here is the new test.

```
        init=inl c' k ->
            inl a1 = a1 c'
            k {
                neutral_elem=-infinityf32,-1
                redo=inl a b -> if fst a > fst b then a else b
                init=inl b k ->
                    inl a1 = a1 b
                    k {
                        neutral_elem=0f32
                        redo=(+)
                        init=inl a -> a1 a .get
                        }
                outit=
                    inl a,b -> a / c', b
                    >> o1 c' .set
                }
```

It is similar to the other one. It is just that I added a `outit` since the output needs to be done on the first thread.

Let me test it out.