``` /// Does arbitrary operations over the innermost dimension using registers. met init_seq w {dim=b,a init} =     inl num_valid = length a     inl items_per_thread, blockDim =         assert (lit_is num_valid) "The inner dimension of the input to this kernel must be known at compile time."         if num_valid <= 1024 then 1, num_valid         else divup num_valid 256, 256
    w.run {
        blockDim
        gridDim={y=min 256 (length b)}
        kernel = cuda
            inl dims = {blockDim gridDim}
            inl grid_for = grid_for {blockDim gridDim}
            grid_for .y b {body=inl {i} ->
                init i <| inl x ->
                    inl block = {
                        store=...
                        store_fst=...
                        map=...
                        redo=...
                        redo'=...
                        broadcast=...
                        load=...
                        }
                    {block} x
                }
        }
```

Let me use this. It is all in some form in the original function. I am going to make it a lot more useful. Maybe at some point I am going to figure out a useful set of abstractions for use in general kernels. `grid_for` is just the tip of the iceberg, there has to be more, but I will need experience before I can get to it. Let me go with tried and true for now.

...Ah, I want to go to bed, but let me put in 1.5 more hour into this. I'll get into it once I start.