Here is the full test. This is quite nice. Adding allocators and asserts in those custom functions was not a problem at all as expected.
Good, good.

Now as tests are concerned, I went through the majority of the work for them. I expected the `seq` function to be one of the toughest and that turned out to be the truth, though for a different reason that I expected initially. I could have done the sampler in a much simpler fashion, but I really like the optimized implementation. It is efficient and maybe I'll get a payoff for doing that at some point.

Let me see what else is there.