Done.
Right now I am not even lazing. It is just that the weight of the effort has finally caught up to me.

I probably removed around 400 lines of code from the module in total. I found a few bugs and fixed them, but I might have caused unknown bugs in turn. Well, probably not. I won't worry about that.

I want to do more than this, but I think I will stop with this work here and call it a job well done.

The allocator version of these functions is something I will consider on a case by case basis. I'll probably do it for redo at least, but I want to avoid doing things that call for me to write type inference functions by hand. Or having to handle multiple map cases everywhere.

I really like how `init_redo_redo` came out. And `seq` for that matter. And the redesigned sampling function, that one is wonderful now.

I really have nothing to be embarassed about with regard to this code.

Let me rest for a bit away from the screen and then it will be time to resume doing what I really should be doing - RNNs.

For the past 5 days my brain has been on kernels. At this point I've pretty much forgotten what I should be doing

Make differentiable plasticity work with PRONG. Test out Zap. Move on to Rudder.

Before that I will have to adjust the learning module so it plays well with the new kernels. I'll do that during the rest of the day.