``` w.CudaKernel.redo_init { ```
Ah, I am calling the wrong kernel. Sometimes I can be too focused on the wrong things. I was really focused on making sure that the indices passed in were correct and I missed something simple like this as a result.

Edit:

```
[|10; 11; 12; 13; 14; 15; 16; 17; 18; 19|]
```

Yeah, now it is right.

I am finally done with this. I am going to take a break here. I think with this I have enough to start work on the learning module, but it good to stop after doing a crucial part of work to let off steam. I really wrote a lot of code today.

Over 200 lines in fact.

http://jmc.stanford.edu/articles/lisp/lisp.pdf
https://fexpr.blogspot.com/2016/08/interpreted-programming-languages.html

Let me read this Lisp stuff.