```     inl map {fwd bck} in s =         inl primal = primals in |> HostTensor.zip         inl out = s.CudaFun.map {map=fwd} primal |> dr s
        {
        out
        bck=inl _ -> join
            inl ins = to_dev_tensor (in,out)
            s.CudaKernel.iter {dim=primal.dim} <| inl i ->
                inl in, out = Struct.map (inl x -> x i) ins
                inl x = bck {in=Struct.map (inl x -> x .get) (primals in); out=primal out .get}
                inl out = adjoint out .get
                Struct.iter2 (inl x -> function
                    | () -> ()
                    | z -> z .set (z .get + out * x)
                    ) x (adjoints in)
        }
```

I am just staring at this and thinking. Is this pleasing? Yeah, it is pleasing all right.

All that maneuvering with multiple `bck`s. that I did is just a waste of time. Let NVCC's optimizer take care of taking out the dead code. I am going to go with this design for the other two functions. That will make it really to do everything that I want. If the adjoint is not there I can easily avoid operating on it.

The above function is extremely understandable compared to what came before. I won't be forced to spend so much time reasoning what the hell does it do like I would with activations. Let me rename it.

With this I think I will take a break here. I am starving.