I am taking a little breather right now.
Essentially, I've done everything I wanted designwise. And I am done with the testing gauntlet, this refactoring will be complete and I will be able to resume work on the RNN. The abstractions I've implemented will allow me to bring out quite a bit of power.

With the `redo` kernel done properly, I'll be able to implement DP as it is in the paper, though it might turn that the version I am using is better. I really hope that the NVCC compiler is doing its job properly with the dead code removal.

Let me give `learning1` a try. Let me see if I can at least get the type errors out of the way today. I won't force myself to get it all done, but there is still a bit left in day so I might as well take the opportunity to do the work that is needed.