``` Training: 0.222503665966006 Testing: 0.174451974073562(9548/9984) Training: 0.0928191526546183 Testing: 0.166122280814586(9640/9984) Training: 0.0371764869421615 Testing: 0.131678517026208(9705/9984) Training: 0.01440738620203 Testing: 0.125657899506934(9732/9984) Training: 0.00574448704232925 Testing: NaN(8997/9216) ```
Ah, it was looking good there for a bit. Damn, so there is a error in there.

It could be anything starting from the sampler. This thing will no doubt keep me busy for a while. I have literally zero clue as to where the problem could be.

Let me try a standard net. It is strange that this error is occuring on the testing phase.

Actually, the first thing I should do is turn on the asserts.