I had a really, really good idea last night. I finally figured out how to do input centering using just the gradients.
Let me explain it. I've long been trying to figure out how to get rid of the PRONG reprojection steps for the input centering part of it. Here is how that could be done.

Let me assume the linear layer is: `(x - c) W + b = x W - c W + b`

It is actually possible to fold `-c W` into `b` so that `d = - c W + b`.

Hence the layer would be `x W + d`.

Now, after a gradient update that linear layer would be: `x (W + dW) - c (W + dW) + b = x (W + dW) - c dW + d`.

That `-c dW` is part of the update for the bias `d`. It is possible to simplify it further.

`-c dW = -c x^T dz`.

As mentioned, it is only a part of the update. The regular gradient for with respect to the loss is just `1 dz`.

So the actual update for the bias `d` rolled into one would be: `-c x^T dz + 1 dz = (-c x^T + 1) dz`.

if one is doing back whitening like in PRONG, then the update would be: `(-c x^T + 1) (R dz)`. This way of formulating it would attain the maximum reuse of available memory and computation. When done on the backward pass instead of the optimization pass, `d` can be more than just weights - it can be some adaptive measure calculated on the fly and backproped through like the weights for differentiable plasticity.

The fact I was not doing input centering was really bothering me up to now and with this the relationship between centering and biases has been elucidated. With this, it would not actually be necessary center the inputs during the forward pass, only the biases need to be updated as shown and all the benefits of centering will be there already.