That idea in the previous commit was a really good way to start this day out.
Now, let me start the day for real. I need to figure out why there are differences in the way the two versions run. It might be just noise due to the different ways the kernels are compiled, but I want to try to ascertain there is nothing wrong in the new way of doing things. The best way to do that is to try and make the two identical.

Let me replace the softmax cross entropy cost with the squared error cost as that one will be more stable. After that I'll try disabling the biases. This is a sanity check. Doing the two should make two version identical with 100% certainty.