No, I can't get the input centering to do anything on Mnist. It first glance it seems sligtly behind not having it, but the difference is negligible enough that it can be considered noise. I'll keep it in for now. Maybe it will work better on RL tasks.
At any rate, I now fully understand the bias reprojection step. This is my accomplishment for the day.

I fixed a bunch of error for the refactoring and investigated the small difference before concluding that it does not really matter. I had to make a bunch of decisions for this.

To be honest, today was pretty disapointing and I feel like I've gone about the day very inefficiently.

But at any rate, I am starting to get into the swing of things again. Before the day is up, let me see if I can get the poker tests up again.