5:05pm. It is running but the time to run a iteration has gone up to 20s. Crap.
5:10pm.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 290 and 710 out of 1000.
The time was 00:00:19.4872816 for: iteration 0
Starting timing for: iteration 1
Winrate is 558 and 442 out of 1000.
The time was 00:00:26.7654071 for: iteration 1
Starting timing for: iteration 2
Winrate is 671 and 329 out of 1000.
The time was 00:00:27.0019431 for: iteration 2
Starting timing for: iteration 3
Winrate is 668 and 332 out of 1000.
The time was 00:00:26.6150228 for: iteration 3
Starting timing for: iteration 4
Winrate is 183 and 817 out of 1000.
The time was 00:00:26.4519641 for: iteration 4
Starting timing for: iteration 5
Winrate is 251 and 749 out of 1000.
The time was 00:00:25.4203654 for: iteration 5
Starting timing for: iteration 6
```

Agh, what a pain. Is this because of inputing centering?

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 303 and 697 out of 1000.
The time was 00:00:12.2074512 for: iteration 0
Starting timing for: iteration 1
Winrate is 569 and 431 out of 1000.
The time was 00:00:17.8517960 for: iteration 1
Starting timing for: iteration 2
Winrate is 697 and 303 out of 1000.
The time was 00:00:16.9979836 for: iteration 2
Starting timing for: iteration 3
Winrate is 710 and 290 out of 1000.
The time was 00:00:17.0398626 for: iteration 3
Starting timing for: iteration 4
Winrate is 714 and 286 out of 1000.
The time was 00:00:16.4003699 for: iteration 4
Starting timing for: iteration 5
Winrate is 463 and 537 out of 1000.
The time was 00:00:14.5967997 for: iteration 5
Starting timing for: iteration 6
Winrate is 247 and 753 out of 1000.
The time was 00:00:12.1272908 for: iteration 6
Starting timing for: iteration 7
Winrate is 238 and 762 out of 1000.
The time was 00:00:11.6529520 for: iteration 7
Starting timing for: iteration 8
Winrate is 184 and 816 out of 1000.
The time was 00:00:10.9433939 for: iteration 8
Starting timing for: iteration 9
Winrate is 160 and 840 out of 1000.
The time was 00:00:10.8826805 for: iteration 9
Starting timing for: iteration 10
Winrate is 119 and 881 out of 1000.
The time was 00:00:10.2368337 for: iteration 10
Starting timing for: iteration 11
Winrate is 137 and 863 out of 1000.
The time was 00:00:12.1166368 for: iteration 11
Starting timing for: iteration 12
Winrate is 155 and 845 out of 1000.
The time was 00:00:11.8351828 for: iteration 12
Starting timing for: iteration 13
Winrate is 109 and 891 out of 1000.
The time was 00:00:11.2027847 for: iteration 13
Starting timing for: iteration 14
```

No it is not.

The truth is I forgot what sort of setting I should be using here. Am I using too large of learning rates or is there something else wrong?

5:20pm.

```
The learning_rate is 2 ** -11.5
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 275 and 725 out of 1000.
The time was 00:00:11.2675268 for: iteration 0
Starting timing for: iteration 1
Winrate is 471 and 529 out of 1000.
The time was 00:00:16.8153080 for: iteration 1
Starting timing for: iteration 2
Winrate is 645 and 355 out of 1000.
The time was 00:00:17.0578261 for: iteration 2
Starting timing for: iteration 3
Winrate is 656 and 344 out of 1000.
The time was 00:00:15.9797247 for: iteration 3
Starting timing for: iteration 4
Winrate is 684 and 316 out of 1000.
The time was 00:00:17.1751685 for: iteration 4
Starting timing for: iteration 5
Winrate is 750 and 250 out of 1000.
The time was 00:00:17.1445800 for: iteration 5
Starting timing for: iteration 6
Winrate is 590 and 410 out of 1000.
The time was 00:00:17.2786323 for: iteration 6
Starting timing for: iteration 7
Winrate is 70 and 930 out of 1000.
The time was 00:00:14.0486968 for: iteration 7
Starting timing for: iteration 8
```

Ah, no, no, no. What is wrong? This is horrible. My redesign not only slowed things down, it made the training unstable.

Let me go to the other branch.