``` The learning_rate is 2 ** -11 ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 393 and 607 out of 1000. The time was 00:00:07.8934168 for: iteration 0 Starting timing for: iteration 1 Winrate is 663 and 337 out of 1000. The time was 00:00:08.4812826 for: iteration 1 Starting timing for: iteration 2 Winrate is 696 and 304 out of 1000. The time was 00:00:08.2792395 for: iteration 2 Starting timing for: iteration 3 Winrate is 698 and 302 out of 1000. The time was 00:00:07.5932541 for: iteration 3 Starting timing for: iteration 4 Winrate is 689 and 311 out of 1000. The time was 00:00:07.8913781 for: iteration 4 Starting timing for: iteration 5 Winrate is 723 and 277 out of 1000. The time was 00:00:07.7452484 for: iteration 5 Starting timing for: iteration 6 Winrate is 721 and 279 out of 1000. The time was 00:00:07.9402972 for: iteration 6 Starting timing for: iteration 7 Winrate is 742 and 258 out of 1000. The time was 00:00:07.5190520 for: iteration 7 Starting timing for: iteration 8 Winrate is 696 and 304 out of 1000. The time was 00:00:07.4952332 for: iteration 8 Starting timing for: iteration 9 Winrate is 707 and 293 out of 1000. The time was 00:00:07.8896532 for: iteration 9 Starting timing for: iteration 10 Winrate is 728 and 272 out of 1000. The time was 00:00:07.3934898 for: iteration 10 Starting timing for: iteration 11 Winrate is 733 and 267 out of 1000. The time was 00:00:07.3774365 for: iteration 11 Starting timing for: iteration 12 Winrate is 730 and 270 out of 1000. The time was 00:00:07.8335425 for: iteration 12 Starting timing for: iteration 13 Winrate is 720 and 280 out of 1000. The time was 00:00:07.7354192 for: iteration 13 Starting timing for: iteration 14 Winrate is 730 and 270 out of 1000. The time was 00:00:07.5441334 for: iteration 14 ```
The old version is definitely not unstable.

I was wrong. There is definitely something wrong with the new version of the library and the error manifests only after some time.

Not only that, but the new thing is almost twice as slow and that is before even putting in the input centering.

Edit:

5:50pm. This thing is just so astonishingly slow with the linear layer. It is literally taking over 25 seconds for a single iteration.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 227 and 773 out of 1000.
The time was 00:00:14.1808406 for: iteration 0
Starting timing for: iteration 1
Winrate is 354 and 646 out of 1000.
The time was 00:00:24.8803538 for: iteration 1
Starting timing for: iteration 2
Winrate is 500 and 500 out of 1000.
The time was 00:00:25.8177545 for: iteration 2
Starting timing for: iteration 3
Winrate is 557 and 443 out of 1000.
The time was 00:00:25.6237864 for: iteration 3
Starting timing for: iteration 4
Winrate is 582 and 418 out of 1000.
The time was 00:00:26.9232931 for: iteration 4
Starting timing for: iteration 5
Winrate is 600 and 400 out of 1000.
The time was 00:00:26.1288809 for: iteration 5
Starting timing for: iteration 6
Winrate is 627 and 373 out of 1000.
The time was 00:00:25.7429125 for: iteration 6
Starting timing for: iteration 7
Winrate is 602 and 398 out of 1000.
The time was 00:00:27.4701335 for: iteration 7
Starting timing for: iteration 8
Winrate is 612 and 388 out of 1000.
The time was 00:00:25.7349511 for: iteration 8
Starting timing for: iteration 9
Winrate is 633 and 367 out of 1000.
The time was 00:00:26.7698258 for: iteration 9
Starting timing for: iteration 10
Winrate is 607 and 393 out of 1000.
The time was 00:00:28.1213561 for: iteration 10
Starting timing for: iteration 11
Winrate is 661 and 339 out of 1000.
The time was 00:00:27.9285309 for: iteration 11
Starting timing for: iteration 12
Winrate is 685 and 315 out of 1000.
The time was 00:00:28.6450222 for: iteration 12
Starting timing for: iteration 13
Winrate is 719 and 281 out of 1000.
The time was 00:00:27.7672840 for: iteration 13
Starting timing for: iteration 14
Winrate is 743 and 257 out of 1000.
The time was 00:00:26.5425786 for: iteration 14
```

Is this because of the biases?