"6:05pm.
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 220 and 780 out of 1000.
The time was 00:00:02.9951255 for: iteration 0
Starting timing for: iteration 1
Winrate is 339 and 661 out of 1000.
The time was 00:00:04.5386578 for: iteration 1
Starting timing for: iteration 2
Winrate is 529 and 471 out of 1000.
The time was 00:00:04.1264192 for: iteration 2
Starting timing for: iteration 3
Winrate is 574 and 426 out of 1000.
The time was 00:00:04.0831799 for: iteration 3
Starting timing for: iteration 4
Winrate is 613 and 387 out of 1000.
The time was 00:00:03.8583001 for: iteration 4
Starting timing for: iteration 5
Winrate is 633 and 367 out of 1000.
The time was 00:00:04.5319363 for: iteration 5
Starting timing for: iteration 6
Winrate is 670 and 330 out of 1000.
The time was 00:00:04.5762317 for: iteration 6
Starting timing for: iteration 7
Winrate is 699 and 301 out of 1000.
The time was 00:00:04.4932273 for: iteration 7
Starting timing for: iteration 8
Winrate is 663 and 337 out of 1000.
The time was 00:00:04.5432454 for: iteration 8
Starting timing for: iteration 9
Winrate is 711 and 289 out of 1000.
The time was 00:00:04.3367054 for: iteration 9
Starting timing for: iteration 10
Winrate is 710 and 290 out of 1000.
The time was 00:00:04.7316926 for: iteration 10
Starting timing for: iteration 11
Winrate is 705 and 295 out of 1000.
The time was 00:00:04.5925239 for: iteration 11
Starting timing for: iteration 12
Winrate is 707 and 293 out of 1000.
The time was 00:00:04.3594535 for: iteration 12
Starting timing for: iteration 13
Winrate is 723 and 277 out of 1000.
The time was 00:00:04.3507630 for: iteration 13
Starting timing for: iteration 14
Winrate is 759 and 241 out of 1000.
The time was 00:00:04.3173901 for: iteration 14
```

This is absoutely unbelievable. The two are identical up to iteration 9. I think the speed difference is due to me setting the grid dimensions too large in the new library. On a whim I've made the default 256 instead of 64. I'll investigate that later along with investigating why the bias functions are so slow.

But I think these blowups can only be because of the cost function. It is there where I should look for clues. I should remove the tanh layer and jack up the learning rates so I can get it to blow up faster.

I think I will save this for tomorrow.

I know that init_seq has been only very losely tested and the same goes for the sampling function. My refactoring was very large and I would have been shocked had bugs like these not snuck in, but still I was hoping I would not have to do this. Hope just leads to despair.

Spending a few days on debugging is a given.

6:15pm. Well, let me do a little extra. I do not feel like stopping now."