"6:20pm. Let me go to the other branch.
```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 395 and 605 out of 1000.
The time was 00:00:05.7218572 for: iteration 0
Starting timing for: iteration 1
Winrate is 647 and 353 out of 1000.
The time was 00:00:06.7749076 for: iteration 1
Starting timing for: iteration 2
Winrate is 691 and 309 out of 1000.
The time was 00:00:06.1535894 for: iteration 2
Starting timing for: iteration 3
Winrate is 647 and 353 out of 1000.
The time was 00:00:03.6535979 for: iteration 3
Starting timing for: iteration 4
Winrate is 679 and 321 out of 1000.
The time was 00:00:03.4679440 for: iteration 4
Starting timing for: iteration 5
Winrate is 716 and 284 out of 1000.
The time was 00:00:03.4184989 for: iteration 5
Starting timing for: iteration 6
Winrate is 678 and 322 out of 1000.
The time was 00:00:03.7242917 for: iteration 6
Starting timing for: iteration 7
Winrate is 702 and 298 out of 1000.
The time was 00:00:03.5330942 for: iteration 7
Starting timing for: iteration 8
```

With this learning rate it should blow up quickly. I mean the new one.

6:25pm.

```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 395 and 605 out of 1000.
The time was 00:00:06.2832067 for: iteration 0
Starting timing for: iteration 1
Winrate is 647 and 353 out of 1000.
The time was 00:00:07.4778850 for: iteration 1
Starting timing for: iteration 2
Winrate is 683 and 317 out of 1000.
The time was 00:00:07.1214375 for: iteration 2
Starting timing for: iteration 3
Winrate is 525 and 475 out of 1000.
The time was 00:00:05.9471678 for: iteration 3
Starting timing for: iteration 4
Winrate is 114 and 886 out of 1000.
The time was 00:00:03.9386813 for: iteration 4
Starting timing for: iteration 5
```

Rather than do it like this, I think I should encounter the same instability on Mnist. I did note the the new one is more unstable, but I thought it was just a fluke. It could not have been."