Ah, I figured out what the problem is just now while I was brushing my teeth.
```
inl redo a b = if fst a <= fst b then a else b
```

When the values get small like in the snipped I gave out earlier comparison between them breaks down. The scan is just a lot of zeroes so the comparison falls into equality and takes the left. Using `<=` might have been a good idea when doing the inclusive scan, but here when the two values are equal, the right one should always be taken as it will be greater than the one on the left.

```
inl redo a b = if fst a < fst b then a else b
```

Doing it like this should fix it. I'll test it out tomorrow.

At this time I'd rather just go to sleep.