Let me try running that thing that I did yesterday night.
```
[|
    [|0.005257684; 0.9931911; 0.000441498; 0.0002269754; 0.0008827905|]
|]

[|1|]

[|
    [|-1.170003; -9.59158; 10.92739; -0.800364; 0.0003000554|]
|]

[|
    [|5.573854E-06; 1.226622E-09; 0.9999683; 8.066552E-06; 1.796437E-05|]
|]

[|2|]

[|
    [|-2.158355; -1.568382; 5.881968; -2.360608; -0.6769392|]
|]

[|
    [|0.0003213743; 0.0005797404; 0.9974226; 0.0002625273; 0.001413781|]
|]

[|2|]

Winrate is 610 and 390 out of 1000.
The time was 00:00:11.7973918 for: iteration 1
```

Yeah, that fixed it.

Let me put back the biases.

Edit:

"Let me put back the biases.

```
Starting timing for: iteration 0
Winrate is 474 and 526 out of 1000.
The time was 00:00:07.4644282 for: iteration 0
Starting timing for: iteration 1
Winrate is 608 and 392 out of 1000.
The time was 00:00:06.6278153 for: iteration 1
```

8:40am. https://arxiv.org/abs/1808.09663
Wasserstein is all you need

```
Starting timing for: iteration 0
Winrate is 502 and 498 out of 1000.
The time was 00:00:03.7950692 for: iteration 0
Starting timing for: iteration 1
Winrate is 610 and 390 out of 1000.
The time was 00:00:03.6061189 for: iteration 1
```

This is without them.

I'll read the NG paper by Amari and this paper, but it seems while I've fixed the issue with the selector, the biases are still an extreme overhead. Honestly, this is completely insane. How is it possible that they are this slow? They are on par or slower than the matrix multiplies.

For now, let me just read the two papers and then I will check it out. I can't belive that kernels are that slow on their own merit. Something is making them slow. `#include <assert.h>` maybe?

Let me read the two papers and then I will start programming."