I had enough of reading papers. Let me start the day for real. I need to investigate why biases are so slow. Those two papers are just filler.
```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 502 and 498 out of 1000.
The time was 00:00:04.0138551 for: iteration 0
Starting timing for: iteration 1
Winrate is 610 and 390 out of 1000.
The time was 00:00:03.9886151 for: iteration 1
Starting timing for: iteration 2
Winrate is 649 and 351 out of 1000.
The time was 00:00:03.8478715 for: iteration 2
Starting timing for: iteration 3
Winrate is 648 and 352 out of 1000.
The time was 00:00:03.7208817 for: iteration 3
Starting timing for: iteration 4
Winrate is 647 and 353 out of 1000.
The time was 00:00:03.5192402 for: iteration 4
```

Here is the thing for 5 iterations without the biases. Let me turn on the asserts and put in the biases.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 474 and 526 out of 1000.
The time was 00:00:14.6942757 for: iteration 0
Starting timing for: iteration 1
Winrate is 608 and 392 out of 1000.
The time was 00:00:13.3316976 for: iteration 1
Starting timing for: iteration 2
Winrate is 633 and 367 out of 1000.
The time was 00:00:13.6342752 for: iteration 2
Starting timing for: iteration 3
Winrate is 604 and 396 out of 1000.
The time was 00:00:13.4080508 for: iteration 3
Starting timing for: iteration 4
Winrate is 626 and 374 out of 1000.
The time was 00:00:12.3521007 for: iteration 4
```

Yeah, asserts wreck it.

```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 474 and 526 out of 1000.
The time was 00:00:07.3525679 for: iteration 0
Starting timing for: iteration 1
Winrate is 608 and 392 out of 1000.
The time was 00:00:06.8666183 for: iteration 1
Starting timing for: iteration 2
Winrate is 633 and 367 out of 1000.
The time was 00:00:06.8000701 for: iteration 2
Starting timing for: iteration 3
Winrate is 604 and 396 out of 1000.
The time was 00:00:06.6527911 for: iteration 3
Starting timing for: iteration 4
Winrate is 626 and 374 out of 1000.
The time was 00:00:06.1739673 for: iteration 4
```

This is without the asserts. Ok, great.

Let me remove asserts entirely. Then I will try commenting out the lines in the HostTensor's apply.