Is it the forward or the backward pass that is slow in the new library? It has to be the backward pass, but let me make sure.
```
The learning_rate is 2 ** -8
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 502 and 498 out of 1000.
The time was 00:00:04.2147251 for: iteration 0
Starting timing for: iteration 1
Winrate is 610 and 390 out of 1000.
The time was 00:00:04.2139321 for: iteration 1
Starting timing for: iteration 2
Winrate is 649 and 351 out of 1000.
The time was 00:00:04.0529416 for: iteration 2
Starting timing for: iteration 3
Winrate is 648 and 352 out of 1000.
The time was 00:00:03.9708871 for: iteration 3
Starting timing for: iteration 4
Winrate is 647 and 353 out of 1000.
The time was 00:00:03.7971744 for: iteration 4
```

It is the backward pass, no doubt about it.

What is so strange about all of this is that for online learning, there is actually no need to do a reduction on the backward pass. The thing should be equivalent to a map.

I am going to have to drill down into this.