``` {blockDim={x=lit 1i64; y=lit 1i64}; gridDim={x=lit 1i64}} ```
Huh, the blockDim's x for this is 1. Well, that would explain the entirey of why the thing is so slow. I never expected the kernel to be sized like this.

```
{a={from=lit 0i64; near_to=lit 1i64}; b={from=lit 0i64; near_to=lit 1i64}}
```

Ah, so `a` and `b` are both 1. Well, I guess then that makes sense. Is the kernel getting called in `redo` or something?

This really should not be happening. Actually `redo` has its own reductions so that should not happen.

I need to track this down. How strange...

Ah, this is probably in the critic.

Let me temporarily turn the stupid thing off then.