``` The learning_rate is 2 ** -11 ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 293 and 707 out of 1000. The time was 00:00:08.0357109 for: iteration 0 Starting timing for: iteration 1 Winrate is 558 and 442 out of 1000. The time was 00:00:10.5971755 for: iteration 1 Starting timing for: iteration 2 Winrate is 657 and 343 out of 1000. The time was 00:00:10.7987723 for: iteration 2 Starting timing for: iteration 3 Winrate is 703 and 297 out of 1000. The time was 00:00:10.7662429 for: iteration 3 Starting timing for: iteration 4 Winrate is 716 and 284 out of 1000. The time was 00:00:10.7315624 for: iteration 4 ```
Yeah, this is how I remember it being. Let me let it run for 15 iterations.

Ah, I am not using biases in naturalize. Let me uncomment them.

It does add some overhead, but nothing like 2.5x anymore.

This is quite satisfactory.

That length thing. I specifically remember having the thought while I was making that particular line that the pattern I was using was dangerous, but rather than doing the right thing, I ignored it. So in order to save a minute then I spent hours tracking it down now. The selector thing was unavoidable without prior knowledge, but this could have gone much better.

But if these were the only two bugs then this is pretty good all in all. I think they were.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 290 and 710 out of 1000.
The time was 00:00:09.0346225 for: iteration 0
Starting timing for: iteration 1
Winrate is 558 and 442 out of 1000.
The time was 00:00:12.1944296 for: iteration 1
Starting timing for: iteration 2
Winrate is 671 and 329 out of 1000.
The time was 00:00:12.3411785 for: iteration 2
Starting timing for: iteration 3
Winrate is 714 and 286 out of 1000.
The time was 00:00:12.2591911 for: iteration 3
Starting timing for: iteration 4
Winrate is 732 and 268 out of 1000.
The time was 00:00:12.0067209 for: iteration 4
Starting timing for: iteration 5
Winrate is 706 and 294 out of 1000.
The time was 00:00:11.3740713 for: iteration 5
Starting timing for: iteration 6
Winrate is 716 and 284 out of 1000.
The time was 00:00:12.1355282 for: iteration 6
Starting timing for: iteration 7
Winrate is 734 and 266 out of 1000.
The time was 00:00:11.1705035 for: iteration 7
Starting timing for: iteration 8
Winrate is 682 and 318 out of 1000.
The time was 00:00:11.1302760 for: iteration 8
Starting timing for: iteration 9
Winrate is 701 and 299 out of 1000.
The time was 00:00:11.9237296 for: iteration 9
Starting timing for: iteration 10
Winrate is 736 and 264 out of 1000.
The time was 00:00:11.4743973 for: iteration 10
Starting timing for: iteration 11
Winrate is 750 and 250 out of 1000.
The time was 00:00:11.6477001 for: iteration 11
Starting timing for: iteration 12
Winrate is 707 and 293 out of 1000.
The time was 00:00:11.8816387 for: iteration 12
Starting timing for: iteration 13
Winrate is 700 and 300 out of 1000.
The time was 00:00:12.1190983 for: iteration 13
Starting timing for: iteration 14
Winrate is 729 and 271 out of 1000.
The time was 00:00:11.7121161 for: iteration 14
```

Exactly as I remember it.

I am really in a good mood now. With this the refactoring is 100% complete. I am at the point where I was over a week ago when I first began this. I am finally done and can make progress in RL again.

I will take a break here. I really deserve it.

This refactoring was not easy. In fact it was quite tedious, but I stuck with it as I always would.

I suppose I could just plug PRONG into the plastic RNN, but what I should do is bring back the char RNN. I do not trust the thing quite yet. In addition to it, I should make a MI-RNN and plug PRONG into it for starters. Pl-RNN is really an online architecture and I do not feel like messing with it on tiny shakespeare as a single pass will take forever.

It is decided - the char RNN will be my next project. I need something to test these things on. Poker is a not a good test for them - I am afraid that Pl-RNN might only be running the feedforward part and I need to absolutely makes sure that this is not the case.

Once that is done I will test the Pl-RNN on poker with PRONG. Once that is done the `make a good agent` part of my plan will be complete. I'll have to try Rudder in order to ensure absolute domination, but even without it, with this I would have enough to try RL seriously.

The last two months were a long, sweltering summer that I spent the entirety in programming, just like the previous two. Things should start getting more interesting from here. I can't leave my fans waiting for too long.